refactor: commons -> utils
