--- src/main/java/dev/rubikon/commons/Event.java
@@ -1,4 +1,0 @@
-package dev.rubikon.commons;
-
-public interface Event {
-}
