+++ src/main/java/dev/rubikon/events/PacketOutgoingEvent.java
@@ -1,6 +1,6 @@
 package dev.rubikon.events;
 
-import dev.rubikon.api.commons.Event;
+import dev.rubikon.utils.Event;
 import net.minecraft.network.packet.Packet;
 
 public class PacketOutgoingEvent implements Event {
