+++ src/main/java/dev/rubikon/events/TickEvent.java
@@ -1,6 +1,6 @@
 package dev.rubikon.events;
 
-import dev.rubikon.commons.Event;
+import dev.rubikon.utils.Event;
 
 /**
  * Invoked every game tick.
