+++ src/main/java/dev/rubikon/utils/Event.java
@@ -1,0 +1,4 @@
+package dev.rubikon.utils;
+
+public interface Event {
+}
