+++ src/main/java/dev/rubikon/utils/Toggleable.java
@@ -1,4 +1,4 @@
-package dev.rubikon.commons;
+package dev.rubikon.utils;
 
 
 /**
